feat: rename base folder
