+++ packages/uui-button-copy-text/lib/UUICopyTextEvent.ts
