+++ packages/uui-button-copy-text/package.json
